requirements.txt
openai
requests
nextcord
python-dotenv
pydantic
setuptools
gspread
oauth2client